requirements.txt
Django==4.1.7
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
kubernetes==26.1.0
kafka-python==2.0.2
django-cleanup==7.0.0
requests==2.28.2
argo-workflows==6.3.0rc2
